requirements.txt
# Core dependencies
httpx>=0.27.0
boto3>=1.34.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-genai>=1.0.0
openai>=1.50.0
python-dotenv>=1.0.0
structlog>=24.1.0
tenacity>=8.2.0
awslambdaric>=2.0.0

# Development dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
mypy>=1.8.0
ruff>=0.2.0
boto3-stubs[dynamodb]>=1.34.0
moto>=5.0.0
